Reorganized classes to better distinguish between build and run.
Specifically, the running code is now a method of the Prebuilt classes,
reflecting the fact that there is no need to involve the parent
Classifier once the Prebuilt(Intersection) is constructed. Relevant
parameters are cloned into the Prebuilts as well; this avoids ambiguity
as to whether parameters affect Prebuilts during build() or run().

Similarly, The IntegratedScorer is now directly produced by the
IntegratedBuilder's finish() call, and has a run() method to run the
classification. This avoids the need for the user to explicitly
construct another class before using the output of finish().

Also fixed the documentation for the Integrated* classes.