Another reorganization to split the build/run code.
This introduces a new set of classes, BasicBuilder and BasicScorer. The former
builds the references and the latter does the classification. The Classifier
class now serves as a convenient wrapper across both classes.

The initial idea in 32a5925 was to do the scoring in a run() method on the
prebuilt instance; however, this meant that running was no longer const w.r.t.
the prebuilt if we wanted to change the parameters. With the changes in this
commit, parameters are set in the BasicScorer, which is then applied to the
prebuilt instance without having to mutate it. This is nicer.

The same logic is applied to the IntegratedBuilder/Scorer, effectively
reverting the changes of 32a5925. Oh well.