Refactor IntegratedBuilder to enable indexed extraction for the scorer. (#15)
This allows the IntegratedScorer to only extract markers from the test matrix
for greater efficiency. However, it involves refactoring the output of the 
IntegratedBuilder so that we report the universe of all markers, in addition
to the various reference-specific fields. We also need to adjust all indices
inside the reference-specific fields so that they index the universe, rather
than indexing the rows of the test matrix.