Propagate custom paralellization scheme to tatami by default.
This is slightly more developer-friendly; it increases the chance that
parallelization is performed consistently in all contexts, albeit at the cost
of possible ODR errors during compilation if the macro isn't set everywhere.

This requires a bump to tatami for the new TATAMI_CUSTOM_PARALLEL signature.
It also requires a name change for the overriding parallelize() function,
to avoid a name clash with the tatami function of the same name
(as macros bypass namespaces) which causes an infinite recursion.