Switch to free functions and Options instead of class methods. (#21)
The former is more intuitive as the (lack of) state is more obvious. It also
gets rid of the option-setting boilerplate and simplifies template deduction.

We use the new knncolle interfaces to perform the neighbor search with
arbitrary algorithms, rather than limiting ourselves to Vptree vs Annoy. The
switch to the new knncolle interfaces also allows us to enable and fix all
compiler warnings, which was previously difficult due to Annoy.

We removed the reference loading code. The exact specification of the file
format does not belong in this library anyway.