Restored default types in user-facing functions for convenience.
Users no longer have to fill out the templated types that can't be deduced from
the arguments, which makes the library a bit more user-friendly. To keep things
a bit more rigorous, the definitions of the default types are stored in
typedefs so that we keep them consistent throughout the library.