Store the expected number of test rows in the Trained* objects.
All classify_* functions now check that the test dataset has the expected
number of rows, based on the arguments provided to the training functions.
This provides an extra safeguard against the use of the wrong test dataset for
a particular Trained object. We also check that all inputs for integrated
training are expecting the same number of rows for the test dataset.

A consequence of this change is that the various function overloads that accept
an intersection now also need to store the expected number of rows in the test
dataset. The current overload has been soft-deprecated and a new overload that
takes this additional argument has been added.